HR/requirements.txt
streamlit
pandas
langchain
langchain-google-genai
python-dotenv
python-docx
pypdf
scikit-learn
openpyxl
tqdm
